Solución al error
