+++ report/__init__.py
@@ -2,3 +2,4 @@
 
 from . import report_purchase_order
 from . import report_invoice_solupaper
+from . import report_logistica_solupaper
